add namedtuple, create cache file in decorators dir
