+++ apps/core/utils.py
@@ -1,7 +1,5 @@
 import logging
-import time
 import urllib
-from functools import wraps
 
 from django.conf import settings
 from django.core.files.base import ContentFile
@@ -9,6 +7,7 @@ from django.template.defaultfilters import slugify as django_slugify
 from rest_framework.response import Response
 
 from apps.core.constants import ALPHABET, STATUS_INFO
+from apps.core.decorators.cache import cache_user
 
 logger = logging.getLogger("django")
 
@@ -94,37 +93,6 @@ def manual_order_model_list(app, admin_site_models_order):
 
     app["models"].sort(key=lambda x: x["name"])
     return app["models"]
-
-
-def decorator_with_args(decorator_to_enhance):
-    """Needs only for other decorator to add some arguments."""
-
-    def decorator_maker(*args, **kwargs):
-        def decorator_wrapper(func):
-            return decorator_to_enhance(func, *args, **kwargs)
-
-        return decorator_wrapper
-
-    return decorator_maker
-
-
-@decorator_with_args
-def cache_user(func, *args, **kwargs):
-    """Cache data for timelimit (in seconds)."""
-    cache_user_dict = dict()
-    timelimit = kwargs["timelimit"]
-
-    @wraps(func)
-    def wrapper(self, request, *args, **kwargs):
-        user = request.user.username
-        current_time = round(time.time() / timelimit)
-        if user in cache_user_dict and current_time == cache_user_dict[user][1]:
-            return cache_user_dict[user][0]
-        result = func(self, request, *args, **kwargs)
-        cache_user_dict[user] = (result, current_time)
-        return result
-
-    return wrapper
 
 
 @cache_user(timelimit=300)
